refactor(financial-occurrence-test): Add test helper to test abstract test
